refactor operation procedures out to OperationProcedures.java
