+++ splice_machine/src/main/java/com/splicemachine/derby/impl/sql/catalog/SpliceSystemProcedures.java
@@ -308,92 +308,7 @@ public class SpliceSystemProcedures extends DefaultSystemProcedureGenerator {
                 .build();
         procedures.add(sessionInfo);
 
-
-        /*
-         * Procedure to get a list of running operations
-         */
-        Procedure runningOperations = Procedure.newBuilder().name("SYSCS_GET_RUNNING_OPERATIONS")
-                .numOutputParams(0)
-                .numResultSets(1)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(runningOperations);
-
-        /*
-         * Procedure to get a list of running operations on the local server
-         */
-        Procedure runningOperationsLocal = Procedure.newBuilder().name("SYSCS_GET_RUNNING_OPERATIONS_LOCAL")
-                .numOutputParams(0)
-                .numResultSets(1)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(runningOperationsLocal);
-
-        /*
-         * Procedure to get a list of progress of running operations
-         */
-        Procedure getProgress = Procedure.newBuilder().name("SYSCS_GET_PROGRESS")
-                .numOutputParams(0)
-                .varchar("drdaTokenFilter",64)
-                .integer("useDrdaToken")
-                .numResultSets(1)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(getProgress);
-
-        /*
-         * Procedure to get a list of progress of running operations on the local server
-         */
-        Procedure getProgressLocal = Procedure.newBuilder().name("SYSCS_GET_PROGRESS_LOCAL")
-                .numOutputParams(0)
-                .varchar("drdaTokenFilter",64)
-                .integer("useDrdaToken")
-                .numResultSets(1)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(getProgressLocal);
-
-
-        /*
-         * Procedure to kill an executing operation
-         */
-        Procedure killOperationLocal = Procedure.newBuilder().name("SYSCS_KILL_OPERATION_LOCAL")
-                .numOutputParams(0)
-                .varchar("uuid",128)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(killOperationLocal);
-
-        /*
-         * Procedure to kill an executing operation
-         */
-        Procedure killOperation = Procedure.newBuilder().name("SYSCS_KILL_OPERATION")
-                .numOutputParams(0)
-                .varchar("uuid",128)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(killOperation);
-
-        /*
-         * Procedure to kill an executing DRDA operation
-         */
-        Procedure killDrdaOperationLocal = Procedure.newBuilder().name("SYSCS_KILL_DRDA_OPERATION_LOCAL")
-                .numOutputParams(0)
-                .varchar("rdbIntTkn",512)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(killDrdaOperationLocal);
-
-        /*
-         * Procedure to kill an executing DRDA operation
-         */
-        Procedure killDrdaOperation = Procedure.newBuilder().name("SYSCS_KILL_DRDA_OPERATION")
-                .numOutputParams(0)
-                .varchar("rdbIntTkn",512)
-                .ownerClass(SpliceAdmin.class.getCanonicalName())
-                .build();
-        procedures.add(killDrdaOperation);
-
+        OperationProcedures.addProcedures(procedures);
 
         /*
          * Procedure to get oldest active transaction id
